Play on click
Due to the AutoPlay policy of chrome, the tab cannot play audio unless the user has crossed the autoplay threshold by clicking on the page. One of the solutions can be to add a Jquery script that plays audio when the user clicks on the page. For example on the coming soon section